server/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
supabase==2.*
python-dotenv==1.*
pydantic==2.*
